Create a public CalculationListener returning a future
